moved routes.go to routes package
