+++ routes/people.go
@@ -1,0 +1,22 @@
+package routes
+
+import (
+	"github.com/go-chi/chi"
+	"github.com/stakwork/sphinx-tribes/handlers"
+)
+
+func PeopleRoutes() chi.Router {
+	r := chi.NewRouter()
+
+	r.Group(func(r chi.Router) {
+		r.Get("/", handlers.GetListedPeople)
+		r.Get("/search", handlers.GetPeopleBySearch)
+		r.Get("/posts", handlers.GetListedPosts)
+		r.Get("/wanteds", handlers.GetListedWanteds)
+		r.Get("/wanteds/assigned/{pubkey}", handlers.GetPersonAssignedWanteds)
+		r.Get("/wanteds/header", handlers.GetWantedsHeader)
+		r.Get("/short", handlers.GetPeopleShortList)
+		r.Get("/offers", handlers.GetListedOffers)
+	})
+	return r
+}
